Add support for a default list of versions.
When compile.sh is called without any parameter, it will look for a
file called ``default-versions.txt`` in the same directory as the
script.
This file should contain the names of the PHP versions you want
compiled and installed on your machine, one per line.
Lines starting with a dash (#) and empty lines are ignored.