Support a "main version".
The first entry in default-versions.txt is used as the main version.
That is, links to it are created in the shared bin/ folder without any
version suffix.
Changing the main version is as simple as swapping the order of lines in
default-versions.txt.

This is an attempt at fixing http://sourceforge.net/p/phpfarm/tickets/6/.