Better patch for default version (cf. e76871d)
This is a better patch for http://sourceforge.net/p/phpfarm/tickets/6/,
following the changes introduced in e76871d.
We now create a symlink to the default version's bin/ dir instead of adding
individual symlinks for each tool.
This is a cleaner approach and also gives access to executables installed
using PEAR in addition to the ones installed by PHP.