Support special tokens in version string.
Makes it possible to pass additional tokens in the version string when
designating a PHP version to install. These tokens will be used to build
a version that includes support for some advanced features.
This is only useful to core developpers though.

Valid tokens:
- "debug"    Compile with debugging symbols (debug version).
- "zts"      Enable thread safety.
- "32bits"   Force the creation of a 32 bits version on a 64 bits machine.
- "gcov"     Enable GCOV code coverage (requires LTP).

Example:
./compile.sh 5.4.0RC1-zts-debug
(the tokens can be passed in any order)