Adds support for the Suhosin patch.
The Suhosin patch will automatically be applied (if available) when the
"suhosin" flag is used in the version string (as in "5.3.9-suhosin").

The scripts were also slightly modified to output all error messages to
stderr instead of stdout.

The README has been updated and now contains an example on how to use
a post-install customization script to install some PEAR extension.