[INCOMPATIBILITY] Use "current" instead of "current-bin".
Rename the symlink (again) from "current-bin" to just "current" and
CHANGE the target of the symlink. Instead of pointing to the "bin/"
directory inside the current version's tree, it only points to the
tree now.
The rationale behing this change is to ease use of php-fpm (which is
installed into sbin/ and not bin/).