Remove support for Suhosin.
The project seems to be dead with no commits over the last year.
Also, the latest version lacks support for the 5.4 & 5.5 branches.

Although direct support for Suhosin has been removed from phpfarm,
it can still be used by adding the appropriate code in a custom
options.sh to fetch and apply the patch manually.